Move entity interface from user bundle into component
