+++ spec/Badger/TagBundle/Security/TaggedEntityVoterSpec.php
@@ -6,7 +6,7 @@ use Badger\Component\Tag\Model\TagInterface;
 use Badger\TagBundle\Security\TaggedEntityVoter;
 use Badger\Component\Tag\Taggable\TaggableInterface;
 use Badger\UserBundle\Entity\User;
-use Badger\UserBundle\Entity\UserInterface;
+use Badger\Component\User\Model\UserInterface;
 use Doctrine\Common\Collections\ArrayCollection;
 use PhpSpec\ObjectBehavior;
 use Prophecy\Argument;
