+++ src/Badger/Bundle/GameBundle/Doctrine/Repository/QuestCompletionRepository.php
@@ -5,7 +5,7 @@ namespace Badger\GameBundle\Doctrine\Repository;
 use Badger\Component\Game\Repository\QuestCompletionRepositoryInterface;
 use Badger\Component\Game\Repository\TagSearchableRepositoryInterface;
 use Badger\UserBundle\Entity\User;
-use Badger\UserBundle\Entity\UserInterface;
+use Badger\Component\User\Model\UserInterface;
 use Doctrine\DBAL\Connection;
 use Doctrine\ORM\EntityRepository;
 use Doctrine\ORM\Query;
