+++ src/Badger/Bundle/GameBundle/Doctrine/Repository/QuestRepository.php
@@ -5,7 +5,7 @@ namespace Badger\GameBundle\Doctrine\Repository;
 use Badger\Component\Game\Repository\QuestRepositoryInterface;
 use Badger\Component\Game\Repository\TagSearchableRepositoryInterface;
 use Badger\Component\Tag\Taggable\TaggableInterface;
-use Badger\UserBundle\Entity\UserInterface;
+use Badger\Component\User\Model\UserInterface;
 use Doctrine\DBAL\Connection;
 use Doctrine\ORM\EntityRepository;
