+++ src/Badger/Bundle/GameBundle/Entity/QuestCompletion.php
@@ -4,7 +4,7 @@ namespace Badger\GameBundle\Entity;
 
 use Badger\Component\Game\Model\QuestCompletionInterface;
 use Badger\Component\Game\Model\QuestInterface;
-use Badger\UserBundle\Entity\UserInterface;
+use Badger\Component\User\Model\UserInterface;
 
 /**
  * @author  Adrien Pétremann <hello@grena.fr>
