+++ src/Badger/Bundle/GameBundle/Factory/UnlockedBadgeFactory.php
@@ -5,7 +5,7 @@ namespace Badger\GameBundle\Factory;
 use Badger\Component\Game\Factory\UnlockedBadgeFactoryInterface;
 use Badger\Component\Game\Model\BadgeInterface;
 use Badger\GameBundle\Entity\UnlockedBadge;
-use Badger\UserBundle\Entity\UserInterface;
+use Badger\Component\User\Model\UserInterface;
 
 /**
  * @author  Adrien Pétremann <hello@grena.fr>
