+++ src/Badger/Component/Game/Factory/UnlockedBadgeFactoryInterface.php
@@ -4,7 +4,7 @@ namespace Badger\Component\Game\Factory;
 
 use Badger\Component\Game\Model\BadgeInterface;
 use Badger\Component\Game\Model\UnlockedBadgeInterface;
-use Badger\UserBundle\Entity\UserInterface;
+use Badger\Component\User\Model\UserInterface;
 
 /**
  * @author  Adrien Pétremann <hello@grena.fr>
