+++ src/Badger/Component/Game/Repository/AdventureStepCompletionRepositoryInterface.php
@@ -3,7 +3,7 @@
 namespace Badger\Component\Game\Repository;
 
 use Badger\Component\Game\Model\AdventureInterface;
-use Badger\UserBundle\Entity\UserInterface;
+use Badger\Component\User\Model\UserInterface;
 
 /**
  * @author  Adrien Pétremann <hello@grena.fr>
