+++ src/Badger/Component/Game/Repository/UnlockedBadgeRepositoryInterface.php
@@ -3,7 +3,7 @@
 namespace Badger\Component\Game\Repository;
 
 use Badger\Component\Game\Model\BadgeInterface;
-use Badger\UserBundle\Entity\UserInterface;
+use Badger\Component\User\Model\UserInterface;
 use Doctrine\Common\Persistence\ObjectRepository;
 use Badger\Component\Tag\Model\TagInterface;
 use Badger\UserBundle\Entity\User;
